\name{predict.SL.dbarts}
\alias{predict.SL.dbarts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.dbarts Objects
}
\description{
This function provides a prediction method for objects of class \code{SL.dbarts}, which are fitted Bayesian Additive Regression Tree (BART) models wrapped within the \code{SuperLearner} framework. It uses the fitted BART model to generate predictions on new data.
}
\usage{
\method{predict}{SL.dbarts}(object, newdata, family, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A fitted object of class \code{SL.dbarts}, typically produced by the \code{SL.dbarts} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{data.frame} of new observations on which to obtain predictions. 
}
  \item{family}{
A \code{list} describing how the prediction should be calculated (currently not used).
}
  \item{\dots}{
Additional arguments (currently not used).
}
}
\value{
A \code{numeric} vector of predicted values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted BART model contained within the \code{SL.dbarts} object.
}
\references{
Chipman, H., George, E., and McCulloch, R. (2009) BART: Bayesian Additive Regression Trees.

The \code{dbarts} package: \url{https://cran.r-project.org/web/packages/dbarts/index.html}
}
\author{
original authors of dbarts: 
Hugh Chipman, Robert McCulloch, Vincent Dorie

Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr
}
\note{
Ensure that the \code{object} provided is a valid \code{SL.dbarts} object, as this function specifically relies on the structure of the BART model fitted by the \code{SL.dbarts} function.
}

\seealso{
\code{\link{SL.dbarts}}, \code{\link[dbarts]{bart}}, \code{\link[SuperLearner]{SuperLearner}}
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (object, newdata, family, ...) 
{
    pred = colMeans(predict(object$object, newdata = newdata))
    return(pred)
  }
}
  \keyword{bart}
  \keyword{regression}
  \keyword{classification}
  \concept{Bayesian Additive Regression Trees}
  \concept{SuperLearner}
