\name{predict.SL.earth_boost}
\alias{predict.SL.earth_boost}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.earth_boost Objects
}
\description{
This function provides a prediction method for objects of class \code{SL.earth_boost}, which are fitted models using boosted Multivariate Adaptive Regression Splines (MARS), wrapped within the \code{SuperLearner} framework. It uses the fitted model to generate predictions on new data.
}
\usage{
\method{predict}{SL.earth_boost}(object, newdata, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A fitted object of class \code{SL.earth_boost}, typically produced by the \code{SL.earth_boost} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{data.frame} of new observations on which to obtain predictions. 
}
  \item{\dots}{
Additional arguments (currently not used).
}
}
\value{
A \code{numeric} vector of predicted values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted MARS model contained within the \code{SL.earth_boost} object.
}
\references{
The \code{SuperLearner} package: \url{https://cran.r-project.org/web/packages/SuperLearner/index.html}

The \code{earth} package: \url{https://cran.r-project.org/web/packages/earth/index.html}
}
\author{
Wrapper from \pkg{SuperLearner} package, written by Eric Polley. 

Extended by  Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr.
}

\seealso{
\code{\link{SL.earth_boost}}, \code{\link[earth]{earth}}
}
\examples{
data(EFV)
result <- SL.earth_boost(Y = EFV[, "VL.1"], X = EFV[, c("efv.0", "metabolic", "weight.0")], 
                         newX = EFV[, c("efv.0", "metabolic", "weight.0")], family = stats::binomial())
predict(object = result$fit, newdata = EFV[, c("efv.0", "metabolic", "weight.0")])
}
\keyword{mars}
\keyword{boosting}
\keyword{regression}
\concept{SuperLearner}
\concept{MARS}
