\name{predict.SL.median}
\alias{predict.SL.median}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.median Objects
}
\description{
This function provides a prediction method for objects of class \code{SL.median}, which are fitted models wrapped within the \code{SuperLearner} framework. It uses the fitted model to generate median-based predictions on new data.
}
\usage{
\method{predict}{SL.median}(object, newdata, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A fitted object of class \code{SL.median}, typically produced by the \code{SL.median} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{data.frame} of new observations on which to obtain predictions. 
}
  \item{\dots}{
Additional arguments (currently not used).
}
}
\value{
A \code{numeric} vector of predicted median values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted model contained within the \code{SL.median} object.
}
\references{
- R Documentation for the \code{matrixStats} package: \url{https://cran.r-project.org/package=matrixStats}
- SuperLearner Package: \url{https://cran.r-project.org/package=SuperLearner}
}
\author{
Original authors of the \code{weightedMedian}: see reference above.

Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr
}
\note{
Ensure that the \code{object} provided is a valid \code{SL.median} object, as this function specifically relies on the structure of the model fitted by the \code{SL.median} function.
}

\seealso{
\code{\link{SL.median}}, \code{\link[SuperLearner]{SuperLearner}}
}
\examples{
data(EFV)
result <- SL.median(Y = EFV[, "VL.1"], X = EFV[, c("efv.0", "metabolic", "weight.0")], 
                    newX = EFV[, c("efv.0", "metabolic", "weight.0")], family = stats::binomial())
predict(object = result$fit, newdata = EFV[, c("efv.0", "metabolic", "weight.0")])
}
\keyword{median}
\keyword{regression}
\keyword{classification}
\concept{SuperLearner}
