\name{predict.SL.randomForest_boost}
\alias{predict.SL.randomForest_boost}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.randomForest_boost objects
}
\description{
This function provides a prediction method for objects of class \code{SL.randomForest_boost}, which are fitted random forest models. It uses the fitted random forest model to generate predictions on new data.
}
\usage{
\method{predict}{SL.randomForest_boost}(object, newdata,  ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A fitted object of class \code{SL.randomForest_boost}, typically produced by the \code{SL.randomForest_boost} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{data.frame} of new observations on which to obtain predictions. 
}
  \item{\dots}{
Additional arguments (currently not used).
}
}
\value{
A \code{numeric} vector of predicted values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted random forest model contained within the \code{SL.randomForest_boost} object.
}
\references{
Breiman, L. (2001), Random Forests, Machine Learning 45(1), 5-32.

Breiman, L (2002), ``Manual On Setting Up, Using, And Understanding Random Forests V3.1'', https://www.stat.berkeley.edu/~breiman/Using_random_forests_V3.1.pdf.

The \code{randomForest} package: \url{https://cran.r-project.org/web/packages/randomForest/index.html}
}
\author{
The function is based on the wrapper from \pkg{SuperLearner} package, written by Eric Polley.  

Extensions by Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr
}
\seealso{
\code{\link[randomForest]{randomForest}}, \code{\link[SuperLearner]{SuperLearner}}
}
\examples{
data(EFV)
result <- SL.randomForest_boost(Y = EFV[, "VL.1"], X = EFV[, c("efv.0", "metabolic", "weight.0")], 
                                     newX = EFV[, c("efv.0", "metabolic", "weight.0")],
                                     family = stats::binomial())
predict(object = result$fit, newdata = EFV[, c("efv.0", "metabolic", "weight.0")])
}
  \keyword{randomforest}
  \keyword{regression}
  \keyword{classification}
  \concept{Random Forest}