\name{predict.SL.step.interaction_boost}
\alias{predict.SL.step.interaction_boost}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.step.interaction_boost objects
}
\description{
This function provides a prediction method for objects of class \code{SL.step.interaction_boost}, which are fitted models using stepwise interactions with boosting, wrapped within the \code{SuperLearner} framework. It uses the fitted model to generate predictions on new data.
}
\usage{
\method{predict}{SL.step.interaction_boost}(object, newdata, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A fitted object of class \code{SL.step.interaction_boost}, typically produced by the \code{SL.step.interaction_boost} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{data.frame} of new observations on which to obtain predictions. 
}
  \item{\dots}{
Additional arguments (currently not used).
}
}
\value{
A \code{numeric} vector of predicted values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted model contained within the \code{SL.step.interaction_boost} object.
}

\author{
  Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr
}

\seealso{
\code{\link{SL.step.interaction_boost}}
}
\examples{
data(EFV)
result <- SL.step.interaction_boost(Y = EFV[, "VL.1"], X = EFV[, c("efv.0", "metabolic", "weight.0")], 
                                    newX = EFV[, c("efv.0", "metabolic", "weight.0")], family = stats::binomial())
predict(object = result$fit, newdata = EFV[, c("efv.0", "metabolic", "weight.0")])
}
\keyword{regression}
\keyword{classification}
\concept{SuperLearner}
