\name{predict.SL.dbarts}
\alias{predict.SL.dbarts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.dbarts objects
}
\description{
This function provides a prediction method for objects of class \code{SL.dbarts}, which are fitted Bayesian Additive Regression Tree (BART) models. It uses the fitted BART model to generate predictions on new data.
}
\usage{
\method{predict}{SL.dbarts}(object, newdata,  ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A fitted object of class \code{SL.dbarts}, typically produced by the \code{SL.dbarts} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{data.frame} of new observations on which to obtain predictions. 
}
  \item{\dots}{
Additional arguments (currently not used).
}
}
\value{
A \code{numeric} vector of predicted values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted BART model contained within the \code{SL.dbarts} object.
}
\references{
Chipman, H., George, E., and McCulloch, R. (2009) \emph{BART: Bayesian Additive Regression Trees}, Annals of Applied Statistics 4(1): 266-298.

The \code{dbarts} package: \url{https://cran.r-project.org/web/packages/dbarts/index.html}
}
\author{
The function is based on the wrapper from package \pkg{tmle}, written by Susan Gruber.

Extensions by Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr.
}
\seealso{
\code{\link{SL.dbarts}}, \code{\link[dbarts]{bart}}
}
\examples{
data(EFV)
result <- SL.dbarts(Y = EFV[, "VL.1"], X = EFV[, c("efv.0", "metabolic", "weight.0")], 
                                     newX = EFV[, c("efv.0", "metabolic", "weight.0")],
                                     family = stats::binomial())
predict(object = result$fit, newdata = EFV[, c("efv.0", "metabolic", "weight.0")])
}
  \keyword{bart}
  \keyword{regression}
  \keyword{classification}
  \concept{Bayesian Additive Regression Trees}
