\name{predict.SL.hal}
\alias{predict.SL.hal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.hal Objects
}
\description{
This function provides a prediction method for objects of class \code{SL.hal}, which are fitted Highly Adaptive Lasso (HAL) models wrapped within the \code{SuperLearner} framework. It uses the fitted HAL model to generate predictions on new data.
}
\usage{
\method{predict}{SL.hal}(object, newdata, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
  A fitted object of class \code{SL.hal}, typically produced by the \code{SL.hal} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{matrix} or \code{data.frame} of new observations on which to obtain predictions. The new data will be automatically converted to a matrix if not already in that format.
}
  \item{\dots}{
  Additional arguments (currently not used).
}
}
\details{
The \code{predict.SL.hal} function extracts the fitted HAL model from the \code{SL.hal} object and generates predictions on the supplied new data. If \code{newdata} is not a matrix, it is automatically converted to ensure compatibility with the HAL model's \code{predict} method.
}
\value{
A \code{numeric} vector of predicted values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted HAL model contained within the \code{SL.hal} object.
}
\references{
Benkeser, D., & van der Laan, M. J. (2016). The highly adaptive lasso estimator.

The \code{hal9001} package: \url{https://github.com/tlverse/hal9001}
}
\author{
  Original authors of the \code{hal9001} package: see reference above.

  Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr
}
\note{
Ensure that the \code{object} provided is a valid \code{SL.hal} object, as this function specifically relies on the structure of the HAL model fitted by the \code{SL.hal} function.
}

\seealso{
\code{\link{SL.hal}}, \code{\link[hal9001]{fit_hal}}, \code{\link[SuperLearner]{SuperLearner}}
}
\examples{
data(EFV)
result <- SL.hal(Y = EFV[, "VL.1"], X = EFV[, c("efv.0", "metabolic", "weight.0")], 
                                     newX = EFV[, c("efv.0", "metabolic", "weight.0")], family = stats::binomial())
predict(object = result$fit, newdata = EFV[, c("efv.0", "metabolic", "weight.0")])
}
\keyword{prediction}
\keyword{regression}
\concept{Highly Adaptive Lasso}
\concept{SuperLearner}
