\name{predict.SL.mgcv}
\alias{predict.SL.mgcv}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prediction Method for SL.mgcv Objects
}
\description{
This function provides a prediction method for objects of class \code{SL.mgcv}, which are fitted generalized additive models (GAM) with smooth terms wrapped within the \code{SuperLearner} framework. It uses the fitted GAM model to generate predictions on new data.
}
\usage{
\method{predict}{SL.mgcv}(object, newdata, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A fitted object of class \code{SL.mgcv}, typically produced by the \code{SL.mgcv} function within the \code{SuperLearner} framework.
}
  \item{newdata}{
  A \code{data.frame} of new observations on which to obtain predictions. 
}
  \item{\dots}{
Additional arguments (currently not used).
}
}
\value{
A \code{numeric} vector of predicted values corresponding to the rows in \code{newdata}. The predictions are generated by the fitted GAM model contained within the \code{SL.mgcv} object.
}
\references{
Wood, S.N., N. Pya and B. Saefken (2016), Smoothing parameter and model selection for general
smooth models (with discussion). Journal of the American Statistical Association 111, 1548-1575
doi:10.1080/01621459.2016.1180986

Wood, S. N. (2020) Inference and computation with generalized additive models and their extensions.
Test 29(2): 307-339. doi:10.1007/s11749020007115

The \code{mgcv} package: \url{https://cran.r-project.org/web/packages/mgcv/index.html}

The \code{SuperLearner} package: \url{https://cran.r-project.org/web/packages/SuperLearner/index.html}
}
\author{
original author of the mgcv package: Simon Wood

Michael Schomaker, Philipp Baumann, Han Bao, Katharina Ring, Christoph Wiederkehr
}
\note{
Ensure that the \code{object} provided is a valid \code{SL.mgcv} object, as this function specifically relies on the structure of the GAM model fitted by the \code{SL.mgcv} function.
}

\seealso{
\code{\link{SL.mgcv}}, \code{\link[mgcv]{gam}}, \code{\link[SuperLearner]{SuperLearner}}
}
\examples{
data(EFV)
result <- SL.mgcv(Y = EFV[, "VL.1"], X = EFV[, c("efv.0", "metabolic", "weight.0")], 
                  newX = EFV[, c("efv.0", "metabolic", "weight.0")], family = stats::binomial())
predict(object = result$fit, newdata = EFV[, c("efv.0", "metabolic", "weight.0")])
}
  \keyword{gam}
  \keyword{regression}
  \keyword{classification}
  \concept{Generalized Additive Models}
  \concept{SuperLearner}